\name{GComponent}
\alias{GComponent}
\title{Base Class. We have
GComponent as parent for GContainer and GWidget.
Here we place QObject and QWidget values, although separating might be better}
\description{
  Base Class. We have GComponent as parent for GContainer
  and GWidget. Here we place QObject and QWidget values,
  although separating might be better
}

