\name{GTable}
\alias{GProgressBar}
\alias{GTable}
\alias{GTimer}
\title{Class for gtable widget}
\description{
  This \code{GTable} class for Qt implements a few
  additional reference methods:

  The main reference methods \code{GTimer} are
  \code{start_timer} and \code{stop_timer}

  For Qt, the Gprogressbar class has the extra reference
  method \code{set_border}. The \code{border} argument has
  been deprecated.
}

