\name{.gwindow.guiWidgetsToolkitQt}
\alias{.ggroup.guiWidgetsToolkitQt}
\alias{.gwindow.guiWidgetsToolkitQt}
\title{toolkit constructor for gwindow}
\usage{
  .gwindow.guiWidgetsToolkitQt(toolkit, title, visible,
    name, width, height, parent, handler, action, ...)

  .ggroup.guiWidgetsToolkitQt(toolkit, horizontal = TRUE,
    spacing = 5, use.scrollwindow = FALSE,
    container = NULL, ...)
}
\description{
  toolkit constructor for gwindow

  toolkit constructor for ggroup
}

